chore: Isola componente toast
